+++ src/pytorch_lightning/overrides/distributed.py
@@ -175,14 +175,10 @@ class IndexBatchSamplerWrapper:
     def __init__(self, sampler: BatchSampler) -> None:
         self.seen_batch_indices: List[List[int]] = []
         self._sampler = sampler
-        self._batch_indices: List[int] = []
-
-
 
     def __iter__(self) -> Iterator[List[int]]:
         self.seen_batch_indices = []
         for batch in self._sampler:
-            self._batch_indices = batch
             self.seen_batch_indices.append(batch)
             yield batch